Create include file for projects wanting to validate externally
In order to allow projects to validate their cat3 files against a cypress ProductTest's expected results, we need to make those libraries available outside of the cypress web app. One approach could be creating a gem which encapsulates this, which could be an approach down the line. Another could be setting up a sepecific web enpoint which would do the validation on an external platform; however, that could make automated testing brittle should the host be unresponsive.

This approach simply allows the validating project to include the `lib/includes_for_external_validation.rb` file and have access to all of the classes and libraries which are used for validation. This commit is not the final product, but rather a stake in the ground, as the minimum set of includes necessary to get the 3rd party project building and able to reference product test classes.