Include Cypress Measure monkeypatch and HDS HQMF module
In order to externally validate using Cypress' "smoking gun" we need to include Cypress' monkeypatch of HealthDataStandard's Measure. This references HQMF which is only defined in the `health-data-standards` gem. However, it is not automagically loaded through said gem. To that end, use em.find_files to load the `hqmf-parser.rb` from `health-data-standards/lib`.