Handle different File duck types for External Validation
In order to allow projects to use Cypress for external validation, we need the ability to handle different IO types that respond to different interfaces than just the one type Cypress has been using so far. Up until now, the object passed in needs to respond to #open; however, on many IO objects, #open is private, and they instead use #read. For that reason, create a helper method that decides which method to use based on the incoming qrda_file object's duck type.