Skip population generation during create callback for External Validation
In order to allow projects to use Cypress for External Validation, we need to allow them to setup a ProductTest using already-existing Record population data, rather than creating it on the fly. For that reason, add a guard clause around the population generation logic that checks to see whether the ProductTest has already been configured. If so, skip population generation during the create callback.